Rename edit-page-component
