+++ src/opening-period/page/EditOpeningPeriodPage.scss
